requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
opencv-python==4.10.0.84
pillow==10.4.0
numpy==1.26.4
timm==1.0.14
python-multipart==0.0.20
mediapipe==0.10.14
# pip install torch==2.4.1 torchvision==0.19.1 -f https://download.pytorch.org/whl/cu121/torch_stable.html